bachman_cortex/requirements.txt
# NOTE: For automatic setup, use ./validate.sh instead.
# This file is for manual installation reference only.
#
# Install order:
#   1. torch + torchvision (requires platform-specific index URL)
#   2. detectron2 (from GitHub, requires --no-build-isolation)
#   3. onnxruntime or onnxruntime-gpu
#   4. pip install -r requirements.txt (this file)
#
# See README.md for full manual setup instructions.

# torch and torchvision are excluded -- they require platform-specific
# index URLs (CPU vs CUDA). Install manually:
#   CPU:  pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu
#   CUDA: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121

# detectron2 is excluded -- not on PyPI, requires:
#   pip install 'git+https://github.com/facebookresearch/detectron2.git' --no-build-isolation

# ONNX Runtime (CPU for local dev, GPU variant on AWS)
onnxruntime>=1.17.0

# Model libraries
insightface>=0.7.3
ultralytics>=8.1.0

# Image/Video processing
opencv-python>=4.9.0
numpy>=1.24.0

# Utilities
Pillow>=10.0
tqdm>=4.65
gdown>=5.0

# Required by 100DOH hand detector (legacy)
easydict
scipy
pyyaml
cython

# Grounding DINO (HuggingFace variant)
transformers>=4.36.0
supervision>=0.18.0
